Rockets fetcher added (subscriber)
